feat: finish paid features responsiveness
