+++ app/styles/pricing/core-features-toggle.module.css
@@ -7,6 +7,9 @@
     font-family: Inter;
     font-style: normal;
     line-height: normal;
+
+    margin: auto;
+    position: relative;
 }
 
 .toggler__container__wrapper {
@@ -18,10 +21,12 @@
 
     margin: auto;
     cursor: pointer;
+
+    margin-top: 84px;
 }
 
 .toggler__item__wrapper__small {
-    width: 260px !important;
+    width: 230px !important;
 }
 
 .toggler__item__wrapper {
@@ -34,7 +39,7 @@
 }
 
 .toggler__item__wrapper__small::before {
-    width: calc(260px + 12px) !important;
+    width: calc(100% + 12px) !important;
 }
 
 .toggler__item__wrapper::before {
@@ -53,7 +58,7 @@
 }
 
 .toggler__item__wrapper__small::after {
-    width: 260px !important;
+    width: 230px !important;
 }
 
 .toggler__item__wrapper::after {
@@ -72,13 +77,14 @@
 }
 
 .toggle__item__container__small {
-    width: 260px !important;
+    width: 230px !important;
 }
 
 .toggle__item__container {
     display: flex;
     flex-direction: column;
     align-items: center;
+    justify-content: center;
 
     background: #000;
 
@@ -135,6 +141,13 @@
 
 .gap__8 {
     gap: 8px !important;
+}
+
+.padding__13 {
+    padding: 13px;
+}
+.padding__none {
+    padding: 0 !important;
 }
 
 .toggler__item__gray {
@@ -266,19 +279,15 @@
     .tooltip > img {
         height: 12px;
     }
-
-    .toggler__container__wrapper{
-        margin: 20px;
-    }
 }
 
 @media (max-width: 468px) {
     .toggler__item__wrapper {
-        max-width: calc(50vw - 20px)
+        max-width: calc(50vw - 20px);
     }
 
     .toggle__item__container {
-        width: calc(50vw - 20px)
+        width: calc(50vw - 20px);
     }
 
     .toggler__item__dark > span:first-of-type {
@@ -295,6 +304,10 @@
         padding: 6px;
     }
 
+    .toggler__container__wrapper {
+        margin: 20px;
+    }
+
     .tooltip {
         right: 60%;
         bottom: 28px;
@@ -308,7 +321,6 @@
         font-size: 0;
     }
 }
-
 
 @media (max-width: 375px) {
     .tooltip {
