+++ app/styles/pricing/feature.module.css
@@ -275,6 +275,7 @@
 
     height: 14px;
     aspect-ratio: 1;
+    z-index: 1;
 }
 
 .tooltip:hover div {
@@ -375,6 +376,8 @@
         display: flex;
         justify-content: center;
         align-items: center;
+
+        border-bottom: 1px solid #2c2c2c;
     }
 
     .mobileFeature > .check {
@@ -388,6 +391,16 @@
         height: 20px;
         width: 20px;
     }
+
+    .expandable > .header .coming-soon-chip {
+        margin-left: 0px;
+    }
+
+    .expandable > .header > span {
+        flex-direction: column;
+        text-align: center;
+    }
+
 }
 
 @media screen and (max-width: 480px) {
@@ -544,6 +557,19 @@
     font-size: 16px;
     font-weight: 400;
     line-height: normal;
+}
+
+.scaleMobile {
+    padding: 12px;
+
+    color: #fff;
+    font-size: 16px;
+    font-weight: 400;
+    line-height: normal;
+
+    display: flex;
+    justify-content: center;
+    gap: 4px;
 }
 
 .seePricing {
