+++ app/styles/pricing/paid-features-toggle.module.css
@@ -1,0 +1,137 @@
+.toggler__container {
+    position: relative;
+    border-radius: 8px;
+    background: #000;
+
+    border: 1px solid #2c2c2c;
+
+    display: flex;
+    gap: 0;
+    font-family: Inter;
+    font-style: normal;
+    line-height: normal;
+
+    cursor: pointer;
+}
+
+.toggler__item__wrapper {
+    position: absolute;
+    padding: 3px;
+    width: 230px;
+
+    height: calc(100% - 6px);
+    top: 3px;
+}
+
+.toggler__item__wrapper::before {
+    width: 100%;
+
+    content: "";
+    position: absolute;
+
+    border-radius: 12px;
+    z-index: 1;
+    background: rgba(255, 153, 51, 0.5);
+    filter: blur(50px);
+    width: calc(100% + 12px);
+    height: calc(100% + 12px);
+
+    top: -5px;
+    left: 3px;
+}
+
+.toggler__item__wrapper::after {
+    width: calc(100% + 2px);
+
+    content: "";
+    position: absolute;
+
+    border: 1px solid #f93;
+    border-radius: 6px;
+
+    z-index: 5;
+
+    top: -1px;
+    left: 2px;
+    height: 100%;
+}
+
+.toggle__item__container {
+    width: 230px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 8px;
+
+    font-size: 16px;
+
+    background: #000;
+
+    height: auto;
+    flex-shrink: 0;
+
+    gap: 10px;
+
+    z-index: 3;
+
+    padding: 14px 12px;
+}
+
+.move__right {
+    right: 8px;
+
+}
+
+.opacity_60 {
+    opacity: 0.6;
+}
+
+.margin_3 {
+    margin: 3px 3px 3px 0;
+}
+
+.tooltip {
+    position: absolute !important;
+    z-index: 5;
+    bottom: 20px;
+    right: 75%;
+}
+
+@media (max-width: 1120px) {
+    .toggler__container {
+        margin: auto;
+    }
+}
+
+@media (max-width: 628px) {
+    .toggler__item__wrapper {
+        width: calc(50vw - 8.434rem);
+    }
+
+    .toggle__item__container {
+        width: calc(50vw - 8.434rem);
+    }
+}
+
+@media (max-width: 480px) {
+    .toggler__item__wrapper {
+        width: calc(50vw - 20px);
+    }
+
+    .toggle__item__container {
+        font-size: 14px;
+        padding: 14px 8px;
+
+        width: calc(50vw - 20px);
+    }
+
+    .toggle__item__container > div > span{
+        font-size: 12px;
+    }
+}
+
+@media (max-width: 375px) {
+    .toggle__item__container{
+        padding: 8px 8px
+    }
+}
