+++ app/styles/pricing/paid-features.module.css
@@ -1,5 +1,6 @@
 .paid__features__container {
     display: flex;
+    flex-direction: row;
     color: white;
     justify-content: space-between;
     align-items: center;
@@ -16,12 +17,12 @@
     margin-bottom: 28px;
 }
 
-.paid__features__container > div > p {
+.paid__features__container > div > span {
     color: #f93;
     font-size: 24px;
     font-style: normal;
     font-weight: 400;
-    line-height: 15px;
+    line-height: 32px;
 }
 
 .paid__feature__note {
@@ -37,6 +38,12 @@
     font-weight: 400;
     line-height: 24px;
 }
+
+.note_mobile {
+    border-top: 0.667px solid #2c2c2c;
+    padding-top: 12px;
+}
+
 
 .pricing__tooltip {
     margin-left: 8px;
@@ -58,6 +65,8 @@
     position: absolute;
     right: -30px;
     top: 28px;
+
+    z-index: 5;
 
     background: #272727;
 }
@@ -88,7 +97,7 @@
     width: calc(100% - 2px);
 
     padding: 12px 12px 0 12px;
-    border: 1px solid #272727
+    border: 1px solid #272727;
 }
 
 .pricing__tooltip__content__container > span {
@@ -131,6 +140,36 @@
     text-align: initial;
 }
 
-.pricing__tooltip__content__container > ul > li:not(:last-child){
+.pricing__tooltip__content__container > ul > li:not(:last-child) {
     margin-bottom: 16px;
+}
+
+@media (max-width: 1120px) {
+    .paid__features__container {
+        flex-direction: column;
+        align-items: start;
+        gap: 72px;
+    }
+
+    .paid__features__container:last-child {
+        margin: auto;
+    }
+}
+
+
+@media (max-width: 700px) {
+    .paid__features__container{
+        margin: 8.438rem 8.438rem 0;
+    }
+}
+
+@media (max-width: 480px) {
+    .paid__features__container{
+        margin: 20px
+    }
+
+    .paid__features__container > div > h1 {
+        font-size: 40px;
+    }
+    
 }
